ai-agent/requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
PyPDF2
python-docx